+++ tfjs-core/src/public/chained_ops/register_all_chained_ops_test.ts
@@ -25,8 +25,8 @@ import {ALL_ENVS, describeWithFlags} from '../../jasmine_util';
 
 const CHAINED_OPS = [
   'add', 'batchNorm', 'broadcastTo', 'concat', 'div', 'divNoNan', 'greater',
-  'less', 'notEqual', 'oneHot', 'pad', 'split', 'square', 'sub', 'tile',
-  'transpose'
+  'less', 'equal', 'notEqual', 'oneHot', 'pad', 'split', 'square', 'sub',
+  'tile', 'transpose'
 ];
 
 describeWithFlags('chained ops', ALL_ENVS, () => {
